remove more code that windows doesnt like
