+++ src/lib.rs
@@ -1,16 +1,11 @@
-pub mod boot_fake_node;
-pub mod boot_real_node;
 pub mod build;
 pub mod build_start_package;
-pub mod chain;
-pub mod connect;
 pub mod dev_ui;
 pub mod inject_message;
 pub mod new;
 pub mod publish;
 pub mod remove_package;
 pub mod reset_cache;
-pub mod run_tests;
 pub mod setup;
 pub mod start_package;
 pub mod update;
